update predicate parsing to reflect rails 6.1 updates
The `WhereClause#ast` method has been changed as of Rails 6.1 to
return only the first predicate if only a single one is present [1].
This means that the returned value from `ast` will not always be
an instance of `Arel::Nodes::And` and we can't therefore always
count on having `children` to iterate on.

This updates the logic to handle the case where a non-And type of node
is returned by `ast`.

[1] https://github.com/rails/rails/pull/39046/commits/699b64ab6322a549a84c7d116fe18cbb522ce64f