convert HomogenousIn node type to In to retain left assignment
A new `Arel::Nodes::HomogeneousIn` node type was introduced in Rails 6.1 [1],
which breaks the logic in `where_with_prefix` as this new type does not allow
for setting its `left` attribute, among other things.

This change converts the Homogeneous node to its `Arel::Nodes::In` counterpart,
which allows the existing logic to assign the sql literal as before. This does
feel like a bit of a quick fix and a refactor may be good at some point to
more soundly accommodate the newly introduced node type.

[1] https://github.com/rails/rails/commit/72fd0bae5948c1169411941aeea6fef4c58f34a9